♻️ refactor: 역할 기반 라우트 구조를 기능 기반 디렉토리 구조로 개편
- 기존 public.py, admin.py 등 역할 중심 라우트 파일을 기능 단위로 분리하여 routes/ 디렉토리 구성
  - public_routes.py, admin_routes.py, friend_routes.py, playlist_routes.py 등으로 나눔
- 템플릿 디렉토리도 기능별로 세분화 (public/, admin/, friends/, user/ 등)
- 기능 중심 구조로 전환함으로써 유지보수성과 파일 탐색성 향상
- Blueprint 등록은 routes/__init__.py 에서 일괄 관리
