--- vibeapp/admin/__init__.py
@@ -1,3 +1,0 @@
-from .routes import admin_bp
-
-__all__ = ['admin_bp']
