--- vibeapp/public/__init__.py
@@ -1,3 +1,0 @@
-from .routes import public_bp
-
-__all__ = ['public_bp']
