+++ vibeapp/routes/admin_routes.py
@@ -7,9 +7,6 @@ admin_bp = Blueprint(
     "admin",
     __name__,
     url_prefix="/admin",
-    
-    #Blueprint가 각자 독립적인 template 디렉토리 사용을 위해 template_folder 옵션 명시
-    template_folder="templates"
 )
 
 @admin_bp.route("/")
@@ -20,10 +17,10 @@ def admin_index():
 @admin_bp.route("/dashboard")
 @admin_required
 def dashboard():
-    return render_template("admin_dashboard.html")
+    return render_template("admin/admin_dashboard.html")
 
 @admin_bp.route("/database")
 @admin_required
 def show_database():
     users = User.query.all()
-    return render_template("admin_database.html", users=users)
+    return render_template("admin/admin_database.html", users=users)
