+++ vibeapp/templates/admin/admin_dashboard.html
