+++ vibeapp/templates/admin/admin_database.html
