+++ vibeapp/templates/user/my_playlists.html
